+++ lib/page-encointer/phases/registering/registerParticipantPanel.dart
@@ -1,19 +1,12 @@
 import 'dart:convert';
 
+import 'package:encointer_wallet/common/components/roundedButton.dart';
+import 'package:encointer_wallet/page/account/txConfirmPage.dart';
+import 'package:encointer_wallet/store/app.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:intl/intl.dart';
-import 'package:encointer_wallet/common/components/infoItem.dart';
-import 'package:encointer_wallet/common/components/roundedButton.dart';
-import 'package:encointer_wallet/common/components/roundedCard.dart';
-import 'package:encointer_wallet/common/consts/settings.dart';
-import 'package:encointer_wallet/page/account/txConfirmPage.dart';
-import 'package:encointer_wallet/service/substrateApi/api.dart';
-import 'package:encointer_wallet/store/app.dart';
-import 'package:encointer_wallet/utils/UI.dart';
-import 'package:encointer_wallet/utils/format.dart';
-import 'package:encointer_wallet/utils/i18n/index.dart';
 
 class RegisterParticipantPanel extends StatefulWidget {
   RegisterParticipantPanel(this.store);
@@ -47,8 +40,6 @@ class _RegisterParticipantPanel extends State<RegisterParticipantPanel> {
       ],
       'onFinish': (BuildContext txPageContext, Map res) {
         Navigator.popUntil(txPageContext, ModalRoute.withName('/'));
-        globalBalanceRefreshKey.currentState.show();
-        globalCeremonyRegistrationRefreshKey.currentState.show();
       }
     };
     Navigator.of(context).pushNamed(TxConfirmPage.route, arguments: args);
@@ -59,25 +50,28 @@ class _RegisterParticipantPanel extends State<RegisterParticipantPanel> {
     // only build dropdown after we have fetched the currency identifiers
     return (store.encointer.participantIndex == null)
         ? CupertinoActivityIndicator()
-        : Column(children: <Widget>[
-            Observer(
-                builder: (_) => (store.encointer.meetupTime == null )
-                ? Container()
-                : Column(children: <Widget>[
-                      Text("Next ceremony will happen at high sun on:"),
-                      Text(DateFormat('yyyy-MM-dd').format(
-                          new DateTime.fromMillisecondsSinceEpoch(
-                              store.encointer.meetupTime)))
-                    ])),
-            Observer(
+        : Column(
+            children: <Widget>[
+              Observer(
+                builder: (_) => (store.encointer.meetupTime == null)
+                    ? Container()
+                    : Column(
+                        children: <Widget>[
+                          Text("Next ceremony will happen at high sun on:"),
+                          Text(DateFormat('yyyy-MM-dd')
+                              .format(new DateTime.fromMillisecondsSinceEpoch(store.encointer.meetupTime)))
+                        ],
+                      ),
+              ),
+              Observer(
                 builder: (_) => store.encointer.participantIndex == 0
-                    ? RoundedButton(
-                        text: "Register Participant",
-                        onPressed: () => _submit())
+                    ? RoundedButton(text: "Register Participant", onPressed: () => _submit())
                     : RoundedButton(
                         text: "Unregister",
                         //for: " + Fmt.currencyIdentifier(store.encointer.chosenCid).toString(),
-                        onPressed: null))
-          ]);
+                        onPressed: null),
+              )
+            ],
+          );
   }
 }